smaller line break

+++ zshrc.py
@@ -100,8 +100,8 @@ alias dockmin="defaults write com.apple.dock tilesize -int 16 && \\
   defaults write com.apple.dock autohide -bool true && \\
   killall Dock"
 
-alias dockleft="defaults write com.apple.dock orientation -string 'left' && \\
-    killall Dock"
+alias dockleft="defaults write com.apple.dock orientation -string \\
+    'left' && killall Dock"
 alias dockdown="defaults write com.apple.dock orientation -string \\
   'bottom' && killall Dock"
